refactor(atoms): merge SfImage
