--- packages/vue/src/components/atoms/SfImage/SfImage.html
@@ -1,78 +1,0 @@
-<div
-  class="sf-image"
-  @mouseover="hoverHandler(true)"
-  @mouseleave="hoverHandler(false)"
-  @click="$emit('click')"
->
-  <transition :name="transition">
-    <div v-if="hasOverlay && loaded" class="sf-image__overlay">
-      <!--@slot slot for overlay content, if is empty the overlay is off -->
-      <slot />
-    </div>
-  </transition>
-
-  <template v-if="lazy">
-    <img
-      v-if="typeof src === 'string'"
-      ref="imgLazy"
-      :alt="alt"
-      :data-src="src"
-      class="sf-image__img"
-    >
-
-    <template v-else-if="src && src.normal">
-      <picture
-        v-if="src.small"
-        ref="imgLazy"
-        class="sf-image__img"
-        :data-iesrc="src.normal.url"
-        :data-alt="src.normal.alt"
-      >
-        <source
-          :srcset="src.small.url"
-          :media="`(max-width: ${pictureBreakpoint - 0.02}px)`"
-        >
-        <source
-          :srcset="src.normal.url"
-          :media="`(min-width: ${pictureBreakpoint}px)`"
-        >
-      </picture>
-
-      <img
-        v-else
-        ref="imgLazy"
-        :alt="src.normal.alt"
-        :data-src="src.normal.url"
-        class="sf-image__img"
-      >
-    </template>
-
-    <img v-else :src="placeholder" alt="No image" class="sf-image__img">
-  </template>
-
-  <template v-else>
-    <img
-      v-if="typeof src === 'string'"
-      :alt="alt"
-      :src="src"
-      class="sf-image__img"
-    >
-    <picture v-else-if="src && src.normal" class="sf-image__img">
-      <source
-        v-if="src.small"
-        :srcset="src.small.url"
-        :media="`(max-width: ${pictureBreakpoint - 0.02}px)`"
-      >
-      <source
-        :srcset="src.normal.url"
-        :media="`(min-width: ${pictureBreakpoint}px)`"
-      >
-      <img
-        :src="src.normal.url"
-        :alt="src.normal.alt"
-        class="sf-image__img"
-      >
-    </picture>
-    <img v-else :src="placeholder" alt="No image" class="sf-image__img">
-  </template>
-</div>
